- refactor login success response models.
