+++ lib/core/shared/cubits/auth_cubit/auth_cubit.dart
@@ -5,9 +5,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:test1/core/shared/cubits/auth_cubit/auth_states.dart';
 import 'package:test1/features/login/data/models/login_request_model.dart';
-import 'package:test1/features/login/data/models/login_success_response_model.dart';
 import 'package:test1/features/sign_up/data/model/sign_up_request.dart';
 
+import '../../../../features/login/data/models/login_success_models/login_success_model.dart';
 import '../../../../features/login/domain/repos/login_repo.dart';
 import '../../../../features/sign_up/data/model/sign_up_response.dart';
 import '../../../../features/sign_up/data/repos/sign_up_repo.dart';
