+++ lib/core/shared/cubits/auth_cubit/auth_states.dart
@@ -3,6 +3,8 @@ import 'dart:developer';
 import 'package:test1/features/login/data/models/login_success_response_model.dart';
 import 'package:test1/features/sign_up/data/model/sign_up_response.dart';
 
+import '../../../../features/login/data/models/login_success_models/login_success_model.dart';
+
 abstract class AuthStates {}
 
 class AuthInitial extends AuthStates {}
@@ -29,7 +31,7 @@ class AuthUserNotExists extends AuthStates {}
 class AuthSignUpSuccess extends AuthStates {
  final SignUpResponse signUpResponse;
 
-  AuthSignUpSuccess(this.signUpResponse);
+ AuthSignUpSuccess(this.signUpResponse);
 }
 class AuthLoginSuccess extends AuthStates {
  final LoginSuccessResponse loginSuccessResponse;
