+++ lib/features/login/domain/repos/login_repo.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import '../../../../core/networking/failure.dart';
-import '../../data/models/login_success_response_model.dart';
+import '../../data/models/login_success_models/login_success_model.dart';
 
 abstract class LoginRepo{
 Future<Either<Failure,LoginSuccessResponse>>login(Map<String,dynamic>data);
