+++ lib/features/login/domain/use_cases/login_use_case.dart
@@ -1,8 +1,9 @@
 import 'package:dartz/dartz.dart';
 import 'package:test1/core/constants/use_cases/use_case.dart';
 import 'package:test1/core/networking/failure.dart';
-import 'package:test1/features/login/data/models/login_success_response_model.dart';
 import 'package:test1/features/login/domain/repos/login_repo.dart';
+
+import '../../data/models/login_success_models/login_success_model.dart';
 
 class LoginUseCase
     implements UseCase<LoginSuccessResponse, Map<String, dynamic>> {
